Support to 10.40+ I no recommend to use client minor to it for new items
TY OTX Team

+++ path_10_11/src/protocolgame.cpp
@@ -1341,7 +1341,7 @@ void ProtocolGame::sendIcons(uint16_t icons)
 {
 	NetworkMessage msg;
 	msg.addByte(0xA2);
-	msg.add<uint16_t>(icons);
+	msg.add<uint32_t>(icons); // TODO: verify compatibility of the new icon range ( 16-31 )
 	writeToOutputBuffer(msg);
 }
